:art: move package manager hook & copy to clipboard hook
